FIX(domain): 매개변수 및 반환 타입 수정
- 구매 금액에 대한 제한이 없기 때문에 막대한 크기의 정수가 들어올 가능성이 있어 long 타입으로 변환

+++ src/main/java/lotto/domain/LottoCalculator.java
@@ -16,8 +16,8 @@ public class LottoCalculator {
         put(7, 1);
     }};
 
-    public int getLottoQuantity(long playerMoney) {
-        return (int) playerMoney / LOTTO_PRICE;
+    public long getLottoQuantity(long playerMoney) {
+        return playerMoney / LOTTO_PRICE;
     }
 
     public double getRate(long rewardMoney, long playerMoney) {
